Implemented parsing of Variation Store Added new operators to DictParser

+++ Adamantium/Adamantium.Fonts/OTF/CFFIndex.cs
@@ -5,7 +5,7 @@ namespace Adamantium.Fonts.OTF
     public struct CFFIndex
     {
         public List<uint> Offsets; // offsets array, last offset is pointing to next table, not to data within current table!!!
-        public ushort Count; // count of real objects in index table, but real offsets count is +1 because last offset points to next table
+        public uint Count; // count of real objects in index table, but real offsets count is +1 because last offset points to next table
         public byte OffsetSize; // size of offset array entry (e.g 2 bytes per each offset)
         public List<byte[]> DataByOffset; // array of data sorted by offsets
         
@@ -13,7 +13,7 @@ namespace Adamantium.Fonts.OTF
         {
             if (Count != 0)
             {
-                return 3 + (Count + 1) * OffsetSize + (Offsets[Count] - 1);
+                return 3 + (Count + 1) * OffsetSize + (Offsets[(int)Count] - 1);
             }
 
             return 2; // according to documentation
